+++ firmware/esp8266/user/widgets/rectangle.h
@@ -6,16 +6,15 @@
 #define RECTANGLE(w) ((rectangle_t*)((w)->priv))
 
 typedef struct {
-    int w, h;
+    uint16_t w, h;
     color_t border_color;
     color_t color;
     color_t altcolor;
     uint8 flash;
     uint8 border;
     uint8 flashcount;
-    int fill:1;
-    int alt:1;
-
+    unsigned int fill:1;
+    unsigned int alt:1;
 } rectangle_t;
 
 #endif
